Relocate ServiceBundler interface to HTTP layer
Resolves #442
