+++ components/app/src/test/java/org/trellisldp/app/SimpleServiceBundler.java
@@ -30,8 +30,8 @@ import org.trellisldp.api.NoopEventService;
 import org.trellisldp.api.NoopMementoService;
 import org.trellisldp.api.NoopNamespaceService;
 import org.trellisldp.api.ResourceService;
-import org.trellisldp.api.ServiceBundler;
 import org.trellisldp.file.FileBinaryService;
+import org.trellisldp.http.core.ServiceBundler;
 import org.trellisldp.io.JenaIOService;
 import org.trellisldp.io.NoopProfileCache;
 import org.trellisldp.triplestore.TriplestoreResourceService;
