+++ components/webdav/src/main/java/org/trellisldp/webdav/impl/WebDAVUtils.java
@@ -41,8 +41,8 @@ import org.trellisldp.api.Metadata;
 import org.trellisldp.api.Resource;
 import org.trellisldp.api.ResourceService;
 import org.trellisldp.api.RuntimeTrellisException;
-import org.trellisldp.api.ServiceBundler;
 import org.trellisldp.api.Session;
+import org.trellisldp.http.core.ServiceBundler;
 import org.trellisldp.http.core.SimpleEvent;
 import org.trellisldp.vocabulary.AS;
 import org.trellisldp.vocabulary.LDP;
