+++ core/http/src/main/java/org/trellisldp/http/TrellisHttpResource.java
@@ -70,8 +70,8 @@ import org.slf4j.Logger;
 import org.trellisldp.api.Metadata;
 import org.trellisldp.api.Resource;
 import org.trellisldp.api.RuntimeTrellisException;
-import org.trellisldp.api.ServiceBundler;
 import org.trellisldp.http.core.PATCH;
+import org.trellisldp.http.core.ServiceBundler;
 import org.trellisldp.http.core.TrellisRequest;
 import org.trellisldp.http.core.Version;
 import org.trellisldp.http.impl.DeleteHandler;
