+++ core/http/src/main/java/org/trellisldp/http/impl/GetHandler.java
@@ -88,8 +88,8 @@ import org.slf4j.Logger;
 import org.trellisldp.api.Binary;
 import org.trellisldp.api.BinaryMetadata;
 import org.trellisldp.api.Resource;
-import org.trellisldp.api.ServiceBundler;
 import org.trellisldp.http.core.Prefer;
+import org.trellisldp.http.core.ServiceBundler;
 import org.trellisldp.http.core.TrellisRequest;
 import org.trellisldp.http.core.Version;
 import org.trellisldp.vocabulary.LDP;
