+++ core/http/src/main/java/org/trellisldp/http/impl/MutatingLdpHandler.java
@@ -51,9 +51,9 @@ import org.trellisldp.api.ConstraintViolation;
 import org.trellisldp.api.Metadata;
 import org.trellisldp.api.Resource;
 import org.trellisldp.api.RuntimeTrellisException;
-import org.trellisldp.api.ServiceBundler;
 import org.trellisldp.api.Session;
 import org.trellisldp.http.core.HttpSession;
+import org.trellisldp.http.core.ServiceBundler;
 import org.trellisldp.http.core.SimpleEvent;
 import org.trellisldp.http.core.TrellisRequest;
 import org.trellisldp.vocabulary.AS;
