+++ core/http/src/main/java/org/trellisldp/http/impl/PatchHandler.java
@@ -65,8 +65,8 @@ import org.trellisldp.api.ConstraintService;
 import org.trellisldp.api.ConstraintViolation;
 import org.trellisldp.api.Resource;
 import org.trellisldp.api.RuntimeTrellisException;
-import org.trellisldp.api.ServiceBundler;
 import org.trellisldp.http.core.Prefer;
+import org.trellisldp.http.core.ServiceBundler;
 import org.trellisldp.http.core.TrellisRequest;
 import org.trellisldp.vocabulary.AS;
 import org.trellisldp.vocabulary.LDP;
