+++ platform/osgi/src/main/java/org/trellisldp/osgi/TrellisServiceBundler.java
@@ -20,7 +20,7 @@ import org.trellisldp.api.EventService;
 import org.trellisldp.api.IOService;
 import org.trellisldp.api.MementoService;
 import org.trellisldp.api.ResourceService;
-import org.trellisldp.api.ServiceBundler;
+import org.trellisldp.http.core.ServiceBundler;
 import org.trellisldp.triplestore.TriplestoreResourceService;
 
 /**
